+++ tools/assign_missing_instruments.php
@@ -87,8 +87,9 @@ function populateVisitLabel($result, $visit_label)
     $battery = new NDB_BVL_Battery;
 
     // select a specific time point (sessionID) for the battery
-    $battery->selectBattery($result['ID']);
-    $timePoint = TimePoint::singleton($result['ID']);
+    $sessionID = new \SessionID(strval($result['ID']));
+    $battery->selectBattery($sessionID);
+    $timePoint = TimePoint::singleton($sessionID);
 
     $DB        = Database::singleton();
     $candidate = Candidate::singleton(new CandID($result['CandID']));
@@ -130,8 +131,8 @@ function populateVisitLabel($result, $visit_label)
 }
 
 if (isset($visit_label)) {
-    $query ="SELECT s.ID, s.subprojectID, s.CandID from session 
-            s LEFT JOIN candidate c USING (CandID) 
+    $query ="SELECT s.ID, s.subprojectID, s.CandID from session
+            s LEFT JOIN candidate c USING (CandID)
             WHERE s.Active='Y'
             AND c.Active='Y' AND s.visit_label=:vl";
     $where = ['vl' => $argv[1]];
@@ -141,8 +142,8 @@ if (isset($visit_label)) {
         populateVisitLabel($result, $visit_label);
     }
 } else if (isset($visit_labels)) {
-    $query   ="SELECT s.ID, s.subprojectID, s.Visit_label, s.CandID from session s 
-            LEFT JOIN candidate c USING (CandID) WHERE s.Active='Y' 
+    $query   ="SELECT s.ID, s.subprojectID, s.Visit_label, s.CandID from session s
+            LEFT JOIN candidate c USING (CandID) WHERE s.Active='Y'
             AND c.Active='Y' AND s.Visit_label NOT LIKE 'Vsup%'";
     $results = $DB->pselect($query, []);
     foreach ($results as $result) {